refactor: 상속 컴포넌트 - Extends 폴더로 위계 재설정
