+++ src/components/@common/Text/Extends/index.tsx
@@ -1,0 +1,9 @@
+import Text from 'components/@common/Text';
+import styled from 'styled-components';
+
+export const EllipsisText = styled(Text)`
+  width: 100%;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
+`;
