+++ src/components/Product/index.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import { ReactComponent as CartIcon } from 'assets/icon/Cart.svg';
 
-import { CartButton } from 'components/@common/Button';
+import { CartButton } from 'components/@common/Button/Extends';
 import Flex from 'components/@common/Flex';
 import { ProductData } from 'types';
 import {
@@ -15,7 +15,8 @@ import {
 import { startCartProductList, setCartProductList } from 'store/cartProductList/actions';
 import { useDispatch } from 'react-redux';
 import Box from 'components/@common/Box';
-import Text, { EllipsisText } from 'components/@common/Text';
+import Text from 'components/@common/Text';
+import { EllipsisText } from 'components/@common/Text/Extends';
 
 const Product = ({ id, thumbnail, name, price }: ProductData) => {
   const dispatch = useDispatch();
